Reorganize litter code
