+++ emergence_lib/src/litter.rs
