+++ emergence_lib/src/ui/production_statistics.rs
@@ -8,8 +8,8 @@ use crate::{
     geometry::Volume,
     items::item_manifest::{Item, ItemManifest},
     light::TotalLight,
+    litter::Litter,
     simulation::{time::InGameTime, weather::CurrentWeather},
-    terrain::litter::Litter,
     units::{item_interaction::UnitInventory, unit_manifest::Unit},
     water::WaterVolume,
     world_gen::WorldGenState,
